plugin/cancel: enable by default
Enable the cancel plugin (with some hackery to prevent cyclic imports)
to be active in every server. If the plugin *is* loaded that takes
precedence, allowing folks to set different timeouts than the default of
5001 ms.

Add several tests and document how to plugin.Done(ctx) in code.
Implement it in *forward* to actually make that return early if we're
only executing that plugin late in the process.

Signed-off-by: Miek Gieben <miek@miek.nl>